refactor(datepicker): re-organize lib structure
