+++ libs/ui-toolkit/datepicker/src/index.ts
@@ -1,1 +1,4 @@
+export * from './lib/adapter/datepicker.adapter';
+export * from './lib/model/datepicker.model';
+export * from './lib/service/datepicker-mode.service';
 export * from './lib/datepicker.module';
