+++ libs/ui-toolkit/datepicker/src/lib/calendar-header/calendar-header.component.scss
