+++ libs/ui-toolkit/datepicker/src/lib/datepicker-content/datepicker-content.component.scss
