+++ libs/ui-toolkit/datepicker/src/lib/datepicker-toggle/datepicker-toggle.component.scss
