+++ libs/ui-toolkit/datepicker/src/lib/model/datepicker.model.ts
